pyproject.toml
[project]
name = "soliddata_mcp_poc"
version = "0.1.0"
description = "CrewAI POC — agent that uses SolidData MCP for text2sql"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[google-genai,tools]>=1.9.3",
    "mcp",
    "httpx",
    "pydantic-settings>=2.0",
    "python-dotenv",
    "snowflake-connector-python>=3.0",
]

[project.scripts]
soliddata_mcp_poc = "soliddata_mcp_poc.main:run"
run_crew = "soliddata_mcp_poc.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
